Associate dogs with owners
Adds a one-to-many relationship between owners and dogs. This also adds
new seed data for users (owners) and updates the routes.

It doesn't change the scoping of the user_dogs_path so that currently
GET user_dogs will return the same results as GET dogs (and not just the
dogs belonging to that user).

Signed-off-by: isabeldepapel <isuchanek@gmail.com>